# [experiment] parallell trie hashing