--- a/trie/proof.go
+++ b/trie/proof.go
@@ -71,7 +71,7 @@ func (t *Trie) Prove(key []byte, fromLevel uint, proofDb ethdb.KeyValueWriter) e
 		// Don't bother checking for errors here since hasher panics
 		// if encoding doesn't work and we're not writing to any database.
 		n, _, _ = hasher.hashChildren(n, nil)
-		hn, _ := hasher.store(n, nil, false)
+		hn, _ := hasher.store(n, nil, false, 0)
 		if hash, ok := hn.(hashNode); ok || i == 0 {
 			// If the node's database encoding is a hash (or is the
 			// root node), it becomes a proof element.
@@ -80,7 +80,7 @@ func (t *Trie) Prove(key []byte, fromLevel uint, proofDb ethdb.KeyValueWriter) e
 			} else {
 				enc, _ := rlp.EncodeToBytes(n)
 				if !ok {
-					hash = hasher.makeHashNode(enc)
+					hash = hasher.makeHashNode(0)
 				}
 				proofDb.Put(hash, enc)
 			}